[IDP-2352] Fix code smell and warnings

+++ src/main/java/it/gov/pagopa/reward/service/reward/RewardCalculatorMediatorServiceImpl.java
@@ -124,8 +124,8 @@ public class RewardCalculatorMediatorServiceImpl extends BaseKafkaBlockingPartit
         if(StringUtils.isEmpty(trx.getChannel())){
             trx.setChannel(RewardConstants.TRX_CHANNEL_RTD);
         }
-        trx.setRuleEngineTopicPartition(CommonUtilities.getHeaderValue(message, KafkaHeaders.RECEIVED_PARTITION));
-        trx.setRuleEngineTopicOffset(CommonUtilities.getHeaderValue(message, KafkaHeaders.OFFSET));
+        trx.setRuleEngineTopicPartition( (Integer) CommonUtilities.getHeaderValue(message, KafkaHeaders.RECEIVED_PARTITION));
+        trx.setRuleEngineTopicOffset( (Long) CommonUtilities.getHeaderValue(message, KafkaHeaders.OFFSET));
     }
 
     @Override
